add a gif of output from pub_sub example
