+++ ROS-Messages-and-Topics/chatter_pkg/CMakeLists.txt
@@ -130,8 +130,8 @@ target_link_libraries(listener ${catkin_LIBRARIES})
 ## Mark executable scripts (Python etc.) for installation
 ## in contrast to setup.py, you can choose the destination
 catkin_install_python(PROGRAMS
-  src/talker.py
-  src/listener.py
+  scripts/talker.py
+  scripts/listener.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
